[BSpline3D] Make a parametric Draft BSpline in 3D from selected vertices
Added ObjectCreation/BSpline3D.FCMacro
This macro makes a parametric bspline from vertices selected in the 3d
view. The BSpline will connect the vertices in the order they have been
selected. The vertices do not have to be on the same shape, part or
body. The BSpline will be fully parametric by means of its Points
property, which will have references to the 3D coordinates of all
selected vertices.