Update Sketch_Constraint_From_Spreadsheet.FCMacro ok wth pad etc...
I changed a piece of code located at the beginning of the "main" procedure. Objective: to make the macro compatible again with pads, pockets, etc. (you select a pad, then a cell in the spreadsheet, and the macro assigns the value of the cell to the Length parameter of the pad).